requirements.txt
Flask
python-dotenv
psycopg2-binary
pydantic
PyJWT
Flask-Cors
gunicorn
waitress 